requirements.txt
lxml
networkx
nltk
matplotlib
wordcloud
tqdm
ipywidgets
widgetsnbextension
pandas-profiling
pathos
